+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-07-23 16:40:36
-// Commit: fix(perf): prevent double tap duplicate action in perf
+// Updated at: 2026-07-23 17:30:05
+// Commit: perf(perf): apply RepaintBoundary on animated perf widgets
 
 class perfComponentState {
-  final String commitRef = '3067';
-  final String lastMessage = 'fix(perf): prevent double tap duplicate action in perf';
-  final DateTime timestamp = DateTime.parse('2026-07-23T16:40:36+05:00');
+  final String commitRef = '3071';
+  final String lastMessage = 'perf(perf): apply RepaintBoundary on animated perf widgets';
+  final DateTime timestamp = DateTime.parse('2026-07-23T17:30:05+05:00');
 }